Identify analogRead() input type and get right pin
This first commit is only for the micro:bit board, but it is to be
extended to the rest of the variants.

The analogRead() argument can be the "ADC Channel" or the pin directly.
Boards without the same type of pin organisation than the Uno (Analog pins
in order after the digital pins) need a look up table to convert a channel
value into the right Pin.