Add connector pins 16 and 17 as analog inputs A6 and A7 Update pin map comments in variant.cpp
Swap microphone (P0.03) with unused pin (P0.23) in the pin map to make it easier to prevent digital pin access to the microphone pin.
(The microphone pin should never be used as a digital output since that could damage the microphone amplifier).
P0.03 is now g_ADigitalPinMap[28]. Since PIN_A0 now points to that entry, and the microphone is only accessed by reading analog pin A0.
This change should be transparent to any existing code.